Relocate tests to colocated homes (#219)
Move three top-level test files to colocated positions per CLAUDE.md
"Definition of done":

- tests/test_promote_admin.py → scripts/dev/test_promote_admin.py
- tests/test_title_case_check.py → scripts/dev/test_title_case_check.py
  (existing 37-line subprocess smoke test renamed to test_title_case_check_smoke.py)
- tests/test_core.py → split into src/core/test_config.py and src/core/test_templating.py

Pure relocation; no test logic changes. Same collected count
(488 passed, 1 skipped) before and after. Updated two README
references to point at the new homes.
